fix(#17): Change the pointcut definition to be more fine grained.
Added "execution" to ensure that we are only capturing the execution of the method and not the method call also.

With the previous definition if an annotated method was called from weaved code it would measure both, meaning we would have an almost duplicated metric, one would include the method call also and the other would be only the execution.