fix(#NA): Test prone to failure due to poor thread handling.
The previous version of this test was prone to failures because the test might end before the threads were executed.

The current implementation doesn't guarantee that the method send method is called, but it does ensure that metric is dropped due to a full queue. This is done by verifying that only one task is submitted to the job queue when putting two metrics